requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
cryptography==3.2.1
decorator==4.4.2
Django==3.1.2
django-rest-knox==4.1.0
djangorestframework==3.12.1
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
tornado==6.0.4
tzlocal==2.1